### 2024-01-18 #6237 @editorialbot
[PRE REVIEW]: nimbleHMC: An R package for Hamiltonian Monte Carlo sampling in nimble
**Submitting author:** @danielturek (<a href="http://orcid.org/0000-0002-1453-1908">Daniel Turek</a>)
**Repository:** https://github.com/nimble-dev/nimbleHMC
**Branch with paper.md** (empty if default branch): joss
**Version:** v0.2.0
