+++ make_package.R
@@ -11,9 +11,9 @@ library(devtools)
 
 setwd('~/github/nimble/nimbleHMC')
 
-document('nimbleHMC')
+devtools::document('nimbleHMC')
 
-system('R CMD BUILD nimbleHMC')
+devtools::build('nimbleHMC')
 
 check('nimbleHMC')
 
@@ -21,7 +21,9 @@ suppressMessages(try(remove.packages('nimbleHMC'), silent = TRUE))
 (tarFiles <- grep('\\.tar\\.gz', list.files(), value = TRUE))
 (lastTarFile <- tarFiles[length(tarFiles)])
 message('installing package version ', gsub('\\.tar\\.gz$', '', lastTarFile))
-system(paste0('R CMD install ', lastTarFile))
+##system(paste0('R CMD install ', lastTarFile))
+
+devtools::install('nimbleHMC')
 
 q('no')
 
