### 2024-01-21 7d36fee
added checks for target nodes of NUTS samplers
