### 2024-01-22 6d65248
avoid deterministic nodes in calcNodes
