### 2024-01-22 d11b6cf
updated HMC tests
