+++ nimbleHMC/tests/testthat/test-HMC.R
@@ -85,13 +85,13 @@ test_that('HMC sampler error messages for transformations with non-constant boun
     Rmodel <- nimbleModel(code, constants = list(c = 1), inits = list(y = 1), buildDerivs = TRUE)
     conf <- configureMCMC(Rmodel, nodes = NULL)
     conf$addSampler('y', 'NUTS')
-    expect_no_error(Rmcmc <- buildMCMC(conf))
+    expect_error(Rmcmc <- buildMCMC(conf))
     ##
     code <- nimbleCode({ x <- 3; y ~ T(dnorm(0, 1), 1, x) })
     Rmodel <- nimbleModel(code, inits = list(y = 2), buildDerivs = TRUE)
     conf <- configureMCMC(Rmodel, nodes = NULL)
     conf$addSampler('y', 'NUTS')
-    expect_no_error(Rmcmc <- buildMCMC(conf))
+    expect_error(Rmcmc <- buildMCMC(conf))
     ##
     code <- nimbleCode({ x ~ dexp(1); y ~ T(dnorm(0, 1), 1, x) })
     Rmodel <- nimbleModel(code, inits = list(x = 3, y = 2), buildDerivs = TRUE)
