### 2024-01-23 27f0161
added checking for non-AD-enabled user distributions
