### 2024-01-23 0ce202f
working to handle environemnts for get
