### 2024-01-24 b026ef8
add tests for new AD functionality: CAR, dyn idx, non-differentiable model pieces
