### 2024-01-24 34fde82
removed check for user-defined distributions
