### 2024-01-24 a04a3d2
bring other tests up to date
