### 2024-01-25 c8f3c50
added back in test for user-dists without AD
