### 2024-01-30 d8c8340
changed is.nan.vec to any_nan
