### 2024-02-01 #6237 @danielturek
@likeajumprope Yes, perhaps (GitHub usernames, taken from the [list of JOSS reviewers](https://reviewers.joss.theoj.org/reviewers)):

- larryshamalama
- thomaspinder
- tbrown122387
- Volkerschmid
