### 2024-02-05 bff8a4b
added print to tests
