### 2024-02-05 7fdbe52
added prints to testing
