### 2024-02-05 2f499a7
Additional checks for HMC samplers (#42)
* added checks for target nodes of NUTS samplers

* avoid deterministic nodes in calcNodes

* updated HMC tests

* minor update to make_package.R

* updated RoxygenNote in DESCRIPTION

* added checking for non-AD-enabled user distributions

* working to handle environemnts for get

* removed check for user-defined distributions

* bring other tests up to date

* temporary remove expected error from distNoAD distribution

* added back in test for user-dists without AD

* trigger testing

* added print to tests

* changed tolerance on test

* added prints
[…]
