+++ nimbleHMC/tests/testthat/test-HMC.R
@@ -31,9 +31,7 @@ test_that('HMC sampler seems to work', {
         set.seed(0)
         samples <- runMCMC(Cmcmc, 100000)
         ##
-        print(abs(as.numeric(apply(samples, 2, mean)) - c(0.4288181, 1.8582433, 3.2853841)))
         expect_true(all(abs(as.numeric(apply(samples, 2, mean)) - c(0.4288181, 1.8582433, 3.2853841)) < 0.017))
-        print(abs(as.numeric(apply(samples, 2, sd)) - c(0.9248042, 1.1964343, 1.3098622)))
         expect_true(all(abs(as.numeric(apply(samples, 2, sd)) - c(0.9248042, 1.1964343, 1.3098622)) < 0.011))
     }
 })
@@ -59,9 +57,7 @@ test_that('HMC sampler on subset of nodes', {
         Cmcmc <- compileNimble(Rmcmc, project = Rmodel)
         set.seed(1)
         samples <- runMCMC(Cmcmc, 100000)
-        print(abs(as.numeric(apply(samples, 2, mean)) - c(0.4288181, 1.8582433, 3.2853841)))
         expect_true(all(abs(as.numeric(apply(samples, 2, mean)) - c(0.4288181, 1.8582433, 3.2853841)) < 0.015))
-        print(abs(as.numeric(apply(samples, 2, sd)) - c(0.9248042, 1.1964343, 1.3098622)))
         expect_true(all(abs(as.numeric(apply(samples, 2, sd)) - c(0.9248042, 1.1964343, 1.3098622)) < 0.013))
     }
 })
