### 2024-02-05 5b6aa3a
Replace `is.nan.vec` with `any_nan` (#46)
* changed is.nan.vec to any_nan

* removed comments about is.nan.vec from test-tuning.R

* trigger testing

* removed prints from testing
