### 2024-02-05 7de254b
added prints to car tests
