+++ nimbleHMC/tests/testthat/test-HMC.R
@@ -598,7 +598,17 @@ test_that('HMC results for CAR match non-HMC', {
     Cmcmc <- compileNimble(Rmcmc, project = Rmodel)
     outHMC <- runMCMC(Cmcmc, niter = 22000, nburnin=2000, thin=20)
 
+   a <- apply(out[,1:6],2,mean)
+   b <- apply(outHMC[,1:6],2,mean)
+   print(a)
+   print(b)
+   print(abs(a-b))
     expect_equal(apply(out[,1:6],2,mean), apply(outHMC[,1:6],2,mean), tolerance = .05)
+   a <- mean(out[,7])
+   b <- mean(outHMC[,7])
+   print(a)
+   print(b)
+   print(abs(a-b))
     expect_equal(mean(out[,7]),mean(outHMC[,7]), tolerance = .15)
 
     expect_equal(apply(out,2,quantile,c(.1,.9)), apply(outHMC,2,quantile,c(.1,.9)), tolerance = 0.15)
