### 2024-02-05 d39b969
add new tests related to new AD functionality (#44)
* add tests for new AD functionality: CAR, dyn idx, non-differentiable model pieces

* trigger testing

* added prints to car tests

* minor change to tolerance

* removed print statements

---------

Co-authored-by: Daniel Turek <danielturek@gmail.com>
