### 2024-02-05 0eb833b
updated package version 0.2.1 and NEWS
