### 2024-02-10 c0c74b4
working on efficiency of hmc_checkTarget
