### 2024-02-10 634164f
working on checkTarget efficiency
