### 2024-02-10 cbddb03
fixed bug using isDiscrete
