### 2024-03-07 #6237 @braniii
Hi @likeajumprope, thx for the invitation. The project sounds very exciting, but unfortunately I have no experience in R and only basic knowledge of HMC. So it probably makes more sense to have someone else review it.
