### 2024-03-08 #6237 @likeajumprope
> Hi @likeajumprope, thx for the invitation. The project sounds very exciting, but unfortunately I have no experience in R and only basic knowledge of HMC. So it probably makes more sense to have someone else review it.

Thanks for letting me know!
