### 2024-03-24 #6237 @arfon
@likeajumprope – just checking to see if any of the automated recommendations from @editorialbot have been checked here too?https://github.com/openjournals/joss-reviews/issues/6237#issuecomment-1898206686
