### 2024-03-24 #6237 @likeajumprope
Hi all, I sent out 4 invitations per email on Friday - hopefully we'll have some replies soon
