### 2024-03-24 #6237 @matt-graham
I received an invite to review this by email. I'd be happy to do so, but just wanted to note I'm not an R expert. I do however know Hamiltonian Monte Carlo well so happy to review on that basis. I'm also currently reviewing another JOSS paper and will be not working over Easter so wouldn't be able to start on reviewing this till a couple of weeks.
