### 2024-04-25 41c782e
Check target efficiency (#47)
* working on efficiency of hmc_checkTarget

* working on checkTarget efficiency

* fixed bug using isDiscrete

* added comments

* updaetd comments
