### 2024-05-13 #6745 @editorialbot
[REVIEW]: nimbleHMC: An R package for Hamiltonian Monte Carlo sampling in nimble
**Submitting author:** @danielturek (<a href="http://orcid.org/0000-0002-1453-1908">Daniel Turek</a>)
**Repository:** https://github.com/nimble-dev/nimbleHMC
**Branch with paper.md** (empty if default branch): joss
**Version:**  0.2.2
