### 2024-05-22 nimble-dev/nimbleHMC#48 @matt-graham
Adding a link to example usage in nimble documentation to README
Raising as part of JOSS review openjournals/joss-reviews/issues/6745

After a bit of searching I was able to find an example of using `nimbleHMC` [in the main nimble user manual](https://r-nimble.org/html_manual/cha-mcmc.html#subsec:HMC-example). While the line

https://github.com/nimble-dev/nimbleHMC/blob/41c782e05532154f35ea8bd91bc1bb153cb9e6fb/README.md?plain=1#L10

does provide a general pointer to the nimble website (and so user manual), it would be good to have a specific link to the example `nimbleHMC` usage somewhere to aid users wanting to get started with using `nimbleHMC`, or including a self-contained example in the README (though appreciate you might not want to have the resulting repetition / need to keep the examples in two places in sync with API changes).
