### 2024-05-22 nimble-dev/nimbleHMC#49 @matt-graham
Adding community guidelines
Raising as part of JOSS review openjournals/joss-reviews/issues/6745

While I see [there are contributing guidelines on the main nimble repository ](https://github.com/nimble-dev/nimble/blob/devel/CONTRIBUTING.md), if contributors to nimbleHMC are advised to follow these for contributions to this specific package this should be explicitly stated, or better some specific contribution guidelines for this repository added.

Similarly it would be helpful to indicate where users should report issues (issue tracker on this repository or main nimbler repository, and when to choose which) and seek support more generally (for example linking to the [nimble-users Google group](https://groups.google.com/forum/?pli=1#!forum/nimble-users).
