### 2024-05-22 nimble-dev/nimbleHMC#50 @matt-graham
Adding installation instructions
Raising as part of JOSS review

openjournals/joss-reviews/issues/6745

It would be useful to add some basic installation instructions to the top-level README similar to (and possibly building on / referring to) [those ](https://github.com/nimble-dev/nimble?tab=readme-ov-file#installation) inthe main nimble repository README
