### 2024-05-22 nimble-dev/nimbleHMC#51 @matt-graham
Linking to API reference documentation on CRAN
Raising as part of JOSS review openjournals/joss-reviews/issues/6745

This would perhaps be obvious to a regular R user, but it took me a while to find that [there is API documentation for the package available on CRAN](https://cran.r-project.org/web/packages/nimbleHMC/nimbleHMC.pdf). It would be useful to provide a pointer to this in the top-level README (and to the top-level CRAN entry https://cran.r-project.org/web/packages/nimbleHMC/).
