### 2024-05-23 nimble-dev/nimbleHMC#48 @danielturek
@matt-graham Great comment.  This is addressed in PR #52.

I opted for adding a link to the user manual, so that we can avoid maintaining examples in multiple different places.
