### 2024-05-23 7a6b78c
added section on Contributing and Support
