### 2024-05-23 97b2b3f
added link to HMC section of user manual (#52)
