### 2024-05-23 fa40e28
updated language in Contributing and Support section
