### 2024-05-23 nimble-dev/nimbleHMC#48 @danielturek
@matt-graham PR #52 has been merged.

I'll leave it to you to review, and close this issue if you agree it's been addressed.
