### 2024-05-23 nimble-dev/nimbleHMC#49 @danielturek
@matt-graham Great suggestions again.

This is being addressed in PR #53
