### 2024-05-23 nimble-dev/nimbleHMC#48 @matt-graham
Thanks looks good to me!
