### 2024-05-23 55d97ed
more detailed installation intructions in README
