### 2024-05-23 3673e4e
Community guidelines (#53)
* added section on Contributing and Support

* updated language in Contributing and Support section
