### 2024-05-23 nimble-dev/nimbleHMC#49 @danielturek
@matt-graham PR #53 has been merged.

I'll leave it to you to review, and close this issue if you agree it's been addressed.
