### 2024-05-23 nimble-dev/nimbleHMC#50 @danielturek
@matt-graham It's a nice idea to give more detailed information about installation.

This is being addressed in PR #54
