### 2024-05-23 0b42fa7
minor updates to samplers avaialble in README
