### 2024-05-23 22d286a
added CRAN API link to README
