### 2024-05-23 nimble-dev/nimbleHMC#51 @danielturek
@matt-graham Yes, that's interesting.  Regular R users would be aware of this, but your experience is helpful to hear.

I've added links to this general package information (including the full package API) in PR #55.
