### 2024-05-23 c2bb7a5
more detailed installation intructions in README (#54)
