### 2024-05-23 nimble-dev/nimbleHMC#49 @matt-graham
Thanks, this looks good to me so closing.
