### 2024-05-23 nimble-dev/nimbleHMC#50 @matt-graham
Thanks, #54 addresses the issue, closing.
