### 2024-05-27 #6745 @danielturek
@crvernon No problem:

- Created review checklist for @matt-graham
- Created review checklist for @larryshamalama
- @matt-graham did (what seems like) a fairly careful review of the [nimbleHMC repository](https://github.com/nimble-dev/nimbleHMC)
- @matt-graham Created four issues for points to address, [48](https://github.com/nimble-dev/nimbleHMC/issues/48), [49](https://github.com/nimble-dev/nimbleHMC/issues/49), [50](https://github.com/nimble-dev/nimbleHMC/issues/50), and [51](https://github.com/nimble-dev/nimbleHMC/issues/51)
- I addressed these issues in PRs [52](https://github.com/nimble-dev/nimbleHMC/pull/52), [53](https://github.com/nimble-dev/nimbleHMC/pull/53), [54](https://github.com/nimble-dev/nimbleHMC/pull/54), and [55](https://github.com/nimble-dev/nimbleHMC/pull/55)
- @matt-graham reviewed these PRs, and closed the corresponding issues
