### 2024-06-01 89c4865
updated hmc_checkTarget to use checkADsupportForDistribution
