### 2024-06-10 nimble-dev/nimbleHMC#59 @matt-graham
Unclear what priors on initial states in latent state sequences in example in JOSS paper
Raising as part of JOSS review openjournals/joss-reviews/issues/6745

In the model definition in the example in the paper

https://github.com/nimble-dev/nimbleHMC/blob/8ca5c0a156d0f446d9dec328496c5c5d74638ce5/joss/paper/paper.md?plain=1#L120-L142

it is unclear to me what the prior distributions assigned on the binary `x[i, first[i]]` variables (that are referenced in the prior on the `x[i, first[i]+1]` variables) will be, that is the first elements in each latent / hidden state sequence `x[i, first[i]:T]`. As the code runs, I assume nimble is doing something sensible here, but the model definition makes it a unclear what this is, and this makes it difficult to try to reproduce this example in other probabilistic programming languages.
