### 2024-06-10 nimble-dev/nimbleHMC#60 @matt-graham
`Detected backwards indexing` warnings when running example in JOSS paper
Raising as part of JOSS review openjournals/joss-reviews/issues/6745

When running the code given in the example in the JOSS paper I get a large series of `Detected backwards indexing` warnings

```
Defining model
  [Warning] Detected backwards indexing in `8:7`. This is likely unintended and will likely not produce valid model code.
  ...
```

when running the snippet

```R
Rmodel <- nimbleModel(
    code,
    constants = list(N = nrow(y), T = ncol(y), f = c(1,2,2,1,1,1,1),
                     first = apply(y, 1, which.max)),
    data = list(y = y),
    inits = list(phi = c(0.5, 0.5), p = 0.5, x = array(1, dim(y))),
    buildDerivs = TRUE)
```

though I think the origin of these warnings is in the specification of the range `first[i]+1:T` in

```R
code <- nimbleCode({
    phi[1] ~ dunif(0, 1)
    phi[2] ~ dunif(0, 1)
    p ~ dunif(0, 1)
    for(i in 1:N) {
        for(t in (first[i]+1):T) {
            x[i,t] ~ dbern(phi[f[t]] * x[i,t-1])
            y[i,t] ~ dbern(p * x[i,t])
        }
    }
})
```

Its unclear if seeing these warnings is expected behaviour or not. From what I can tell it corresponds to the values of `i` for which `first[i] == 7`, that is the rows of `y` for which the first column (reading from left) to have 1 entry is the last (7th) column. If this behaviour is expected it would probably be worth adding a note to that effect to the text in the paper.

The rationale for iterating over `t` in `first[i]+1:T` rather than `1:T` would also benefit from some explanation as it's not entirely clear to me from the description why we are doing so.
