### 2024-06-14 ccd70d7
updated version number to 0.2.2
