### 2024-06-14 8459004
removed unused variables old_stepsize and old_epsilon
