### 2024-06-14 1e336ea
updated hmc_checkTarget to use checkADsupportForDistribution (#57)
* updated hmc_checkTarget to use checkADsupportForDistribution

* removed unnecessary deparse

* fixed minor typo

* trigger testing
