+++ nimbleHMC/DESCRIPTION
@@ -15,3 +15,4 @@ Copyright: See COPYRIGHTS file.
 Encoding: UTF-8
 LazyData: false
 RoxygenNote: 7.3.1
+
