### 2024-06-16 nimble-dev/nimbleHMC#60 @danielturek
Addressed in #63
