### 2024-06-16 0e05f9f
added explanation of x[i, first[i]] to manuscript
