### 2024-06-16 18de386
clarify sampler assignment in text
