### 2024-06-16 nimble-dev/nimbleHMC#61 @danielturek
Addressed in #65
