### 2024-06-16 nimble-dev/nimbleHMC#59 @danielturek
@matt-graham This is addressed in #64, which clarifies the state variable value for each first observation period.  Could you take a look at #64 and let me know if this resolves this issue?
