### 2024-06-16 nimble-dev/nimbleHMC#60 @danielturek
@matt-graham This has been addressed in #63, which removes any backwards indexing, and the example no longer produces this warning.  The inferential results are unchanged, but these individuals (first sighted on the final occasion) are no longer included.  The text is also updated to explain this.  Please let me know if this resolves the issue.
