### 2024-06-16 nimble-dev/nimbleHMC#61 @danielturek
@matt-graham Thanks for pointing this out, this was fascinating to read!

I updated the manuscript.  I chose to *not* attempt a review of capabilities of other packages, but rather to remove any claims of uniqueness (I thought this was safer).  The manuscript now only describes the functionality of the package, rather than making claims of uniqueness among software.  Could you take a look at the updated text in #65, and see if this resolves the issue?
