### 2024-06-17 nimble-dev/nimbleHMC#60 @matt-graham
Thanks @danielturek for the explanation and fixes. The changes to the paper look good to me and confirming I no longer see the warnings
