### 2024-06-17 nimble-dev/nimbleHMC#60 @matt-graham
Sorry just realised that #63 is not yet merged in so possibly should have waited till it was before closing!
