### 2024-06-17 nimble-dev/nimbleHMC#59 @matt-graham
The additional clarifications looks great to me @danielturek - thanks!
