### 2024-06-17 nimble-dev/nimbleHMC#61 @matt-graham
Thanks @danielturek - the rephrasing to remove claim of uniqueness looks good to me.
