### 2024-06-17 d9a5973
removed individuals only seen on final occasion (#63)
