### 2024-06-17 nimble-dev/nimbleHMC#60 @danielturek
Thanks @matt-graham for the helpful review!

#63 is now merged.
