### 2024-06-17 4639598
added explanation of x[i, first[i]] to manuscript (#64)
