### 2024-06-17 6714a12
clarify sampler assignment in text (#65)
