### 2024-06-17 nimble-dev/nimbleHMC#61 @danielturek
Thanks one more time, @matt-graham .  I really appreciate your helpful comments.

#65 is now merged.  Closing this issue.
