### 2024-06-18 03348b0
Defer initEpsilon() to start of next iteration
