### 2024-06-19 #6745 @larryshamalama
@crvernon @danielturek Sincere apologies for the delay. We had some family issues that came up and a lot of my work took significant delays...

Overall, everything looks great: the paper, the repository (with respect to all items in the checklist) and, most importantly, the new HMC sampler to `nimble`. To prevent further delays, I'm very happy to recommend this to be accepted. I have some very minor suggestions; I can raise them as issues or PRs in the repository
