### 2024-06-19 nimble-dev/nimbleHMC#67 @larryshamalama
Change example in documentation?
https://github.com/nimble-dev/nimbleHMC/blob/1e336eace85a1ffa0582f8396eab59cdf40fe0a7/nimbleHMC/R/HMC_samplers.R#L30-L52

(same example as in the [reference manual](https://cran.r-project.org/web/packages/nimbleHMC/nimbleHMC.pdf))

I tried this example at first and, while it runs well, I was a bit confused as to what it is estimating. While the posterior plots seem confusing at first, the posterior draws have little stochasticity, since the data does not have any randomness as well. Would it be better to use a more "illustrative" generated data? E.g.

```{r}
N <- 1000
x <- rnorm(N)
y <- 1 + 2 * x + rnorm(N)

# nimble model

constants <- list(N = N, x = x)
data <- list(y = y)

# etc.
```

CC https://github.com/openjournals/joss-reviews/issues/6745
