### 2024-06-19 92824a1
Capitalization of some PPL frameworks
