### 2024-06-21 b34abba
updated HMC roxygen example model
