+++ nimbleHMC/R/HMC_samplers.R
@@ -38,9 +38,12 @@
 ## #'     }
 ## #' })
 ## #' 
-## #' N <- 10
-## #' constants <- list(N = N, x = 1:N)
-## #' data <- list(y = 1:N)
+## #' set.seed(0)
+## #' N <- 100
+## #' x <- rnorm(N)
+## #' y <- 1 + 0.3*x + rnorm(N)
+## #' constants <- list(N = N, x = x)
+## #' data <- list(y = y)
 ## #' inits <- list(b0 = 1, b1 = 0.1, sigma = 1)
 ## #' 
 ## #' Rmodel <- nimbleModel(code, constants, data, inits, buildDerivs = TRUE)
@@ -307,9 +310,12 @@ hmc_setWarmup <- nimbleFunction(
 #'     }
 #' })
 #' 
-#' N <- 10
-#' constants <- list(N = N, x = 1:N)
-#' data <- list(y = 1:N)
+#' set.seed(0)
+#' N <- 100
+#' x <- rnorm(N)
+#' y <- 1 + 0.3*x + rnorm(N)
+#' constants <- list(N = N, x = x)
+#' data <- list(y = y)
 #' inits <- list(b0 = 1, b1 = 0.1, sigma = 1)
 #' 
 #' Rmodel <- nimbleModel(code, constants, data, inits, buildDerivs = TRUE)
@@ -772,9 +778,12 @@ treebranchNL_NUTS <- nimbleList(p_beg = double(1), p_end = double(1), rho = doub
 #'     }
 #' })
 #' 
-#' N <- 10
-#' constants <- list(N = N, x = 1:N)
-#' data <- list(y = 1:N)
+#' set.seed(0)
+#' N <- 100
+#' x <- rnorm(N)
+#' y <- 1 + 0.3*x + rnorm(N)
+#' constants <- list(N = N, x = x)
+#' data <- list(y = y)
 #' inits <- list(b0 = 1, b1 = 0.1, sigma = 1)
 #' 
 #' Rmodel <- nimbleModel(code, constants, data, inits, buildDerivs = TRUE)
