### 2024-06-21 nimble-dev/nimbleHMC#67 @danielturek
@larryshamalama Agreed, the example in the roxygen was not ideal.  I've just changed the HMC example to something similar to your suggestion, the relevant code:

```
set.seed(0)
N <- 100
x <- rnorm(N)
y <- 1 + 0.3*x + rnorm(N)
constants <- list(N = N, x = x)
data <- list(y = y)
```

This change is made on the `master` branch of the `nimbleHMC` repository, and also merged into the `joss` branch for this review.

Thanks for the helpful suggestion.  I'm going ahead and closing this issue.
