### 2024-06-21 #6745 @danielturek
@crvernon I've addressed the additional comments, with details below.

And thank you to @larryshamalama for the additional review.

- Fixed [issue 67](https://github.com/nimble-dev/nimbleHMC/issues/67), by updating the examples in the roxygen in `HMC_samplers.R`.  The examples are now more realistic following the suggestions of @larryshamalama .
- Merged [PR 68](https://github.com/nimble-dev/nimbleHMC/pull/68) which contained some capitalization corrections for PPL frameworks, thanks for making this PR @larryshamalama

@crvernon Please let me know if there's anything further I can / should do.
