### 2024-06-28 c793a4e
added package anchors to links in roxygen
