### 2024-07-02 #6745 @likeajumprope
> Paper is not ready for acceptance yet, the archive is missing

Exactly, we need a link to an archive containing the latest (reviewed) version of the software (Zenodo/Figshare)
