### 2024-07-04 #6745 @danielturek
@likeajumprope Sounds good.

I just released a new version of the software including the latest changes from this review (version 0.2.2), and created a zenodo repository with that version of the code:
[https://doi.org/10.5281/zenodo.12658544](https://doi.org/10.5281/zenodo.12658544)

Let me know if there's anything else I can do.
