### 2024-07-05 #6745 @likeajumprope
> ```
> Reference check summary (note 'MISSING' DOIs are suggestions that need verification):
>
> OK DOIs
>
> - 10.21105/joss.03844 is OK
> - 10.1080/10618600.2016.1172487 is OK
> - 10.18637/jss.v076.i01 is OK
> - 10.1214/aos/1056562461 is OK
> - 10.1214/16-BA1008 is OK
> - 10.1002/ece3.6053 is OK
> - 10.1080/10618600.2013.791193 is OK
> - 10.2307/2937171 is OK
> - 10.1007/s10651-016-0353-z is OK
>
> MISSING DOIs
>
> - No DOI given, and none found for title: Handbook of markov chain monte carlo
> - No DOI given, and none found for title: JAGS: A program for analysis of Bayesian graphical...
> - No DOI given, and none found for title: PyMC: Bayesian stochastic modelling in python
> - No DOI given, and none found for title: Stan modeling language users guide and reference m...
> - No DOI given, and none found for title: The metropolis—Hastings algorithm
> - No DOI given, and none found for title: Conjugate likelihood distributions
> - No DOI given, and none found for title: The No-U-Turn sampler: adaptively setting path len...
> - No DOI given, and none found for title: Discrete Latent Variable Models
> - No DOI given, and none found for title: Elliptical slice sampling
> - No DOI given, and none found for title: mra: Mark-Recapture Analysis
> - No DOI given, and none found for title: WinBUGS-a Bayesian modelling framework: concepts, ...
> - No DOI given, and none found for title: Composable Effects for Flexible and Accelerated Pr...
> - No DOI given, and none found for title: Deep learning with tensorflow: A review
>
> INVALID DOIs
>
> - None
> ```

The only issue that I now still have to remark is that some of the papers don't have dois. Could you add those @danielturek (sorry to be pedantic here)? As soon as it's done I'll regenerate the Pdf and we're good to go.
