+++ joss/paper/paper.bib
@@ -120,7 +120,8 @@
   volume={9},
   pages={425--452},
   year={2022},
-  publisher={Annual Reviews}
+  publisher={Annual Reviews},
+  doi={10.1146/annurev-statistics-040220-091910}
 }
 
 @article{turek2017automated,
@@ -206,14 +207,16 @@
   volume={10},
   pages={325--337},
   year={2000},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1023/A:1008929526011}
 }
 
 @article{phan2019composable,
   title={Composable Effects for Flexible and Accelerated Probabilistic Programming in NumPyro},
   author={Phan, Du and Pradhan, Neeraj and Jankowiak, Martin},
   journal={arXiv preprint arXiv:1912.11554},
-  year={2019}
+  year={2019},
+  doi={10.48550/arXiv.1912.11554}
 }
 
 @article{pang2020deep,
@@ -224,7 +227,8 @@
   number={2},
   pages={227--248},
   year={2020},
-  publisher={SAGE Publications Sage CA: Los Angeles, CA}
+  publisher={SAGE Publications Sage CA: Los Angeles, CA},
+  doi={10.3102/1076998619872761}
 }
 
 
