### 2024-07-05 #6745 @danielturek
@likeajumprope Thanks for the suggestion.

I took a careful look at all the missing dois, and found a few which I added in [a0f1be6](https://github.com/nimble-dev/nimbleHMC/commit/a0f1be6b58acac73791ce7755e3dfaab994f029d).

I don't believe there are any more, but please let me know if I'm overlooking something.
