### 2024-07-08 #6745 @crvernon
:wave: @danielturek - please address the following before I can move forward with accepting your submission for publication:

- [ ] Ensure that the author names and order in your Zenodo archive match those represented in your paper exactly.  You can edit the Zenodo archive metadata to do this and do not have to conduct another release.
- [ ] Ensure that the title in your Zenodo archive match that represented in your paper exactly.  You can edit the Zenodo archive metadata to do this and do not have to conduct another release.
- [ ] Please ensure that capitalization in maintained in your references where appropriate.  For example, LINE 166 states "An r package for studying..." for which "r" should be capitalized as "R".  You can achieve this by using curly brackets around formatting you wish to keep in the bib file.  Please check all references for this.

Once the above are complete, I will move forward with accepting this submission for publication.
