### 2024-07-09 f4deed7
updated capitalization of de Valpine 2022 reference
