+++ joss/paper/paper.bib
@@ -14,7 +14,7 @@
 }
 
 @article{de2022comparemcmcs,
-  title={compareMCMCs: An R package for studying MCMC efficiency},
+  title={compareMCMCs: An {R} package for studying MCMC efficiency},
   author={{de Valpine}, Perry and Paganin, Sally and Turek, Daniel},
   journal={Journal of Open Source Software},
   volume={7},
