### 2024-07-09 #6745 @danielturek
@crvernon I made minor edits to the metadata of the [Zenodo archive](https://zenodo.org/records/12658544).

- Confirming that the author names and order now exactly match those in the manuscript.
- Confirming that the title now exactly matches that of the manuscript.

I also corrected the reference you mentioned, and checked the others.  It now appears correctly in the article proof.

Please let me know if these changes look ok, and thanks for the careful review.
